better timing calc in 18-framerate

+++ 18-framerate/c/src/main.c
@@ -21,9 +21,9 @@ int main(int argc, char *argv[]) {
 
   while (1==1) {
     neo_console_set_cursor_pos(0,0);
-    unsigned long t = neo_system_timer(), fc = neo_graphics_frame_count();
+    unsigned long long t = neo_system_timer(), fc = neo_graphics_frame_count();
     unsigned long long fps = (100000 * fc)/t;
-    printf("timer: %lu -- frame: %lu -- 100k x fps: %llu",t,fc,fps);
+    printf("timer: %llu -- frame: %llu -- 100k x fps: %llu",t,fc,fps);
     while (fc == neo_graphics_frame_count()) /* wait */ ;
     }
   }
